My current example had a race issue, where concurrent transfers could override each other, resulting in the balance on the accounts to be out of sync with the transfer ledger, and not reconcile.
I was aware of that issue, and since it can be database dependent how you
address it, I hadn't implemented a way to solve it, as I was more focused on
describing the general DDD concepts and structure. But as some have commented,
this might make the example confusing or misleading, where people might not know
how to structure within DDD a solution to this problem.

So I updated my example to solve this issue in the most commonly used DDD
pattern, which is something normally called Event Sourcing.

This approach is eventually consistent, which means there's no read-afer-write
guarantees, and it has its pros/cons. What happens is that the domain model and
domain services when used to change the state of the application, will return a
domain event that describes the valid change. Those events are then eventually
handled by the application service and passed to the repository to be applied
and committed to the actual data-store maintaining the app state.

Everything will be eventually consistent, but the check for debiting only if
there is enough balance becomes best effort, and not an absolute guarantee, thus
loosening the requirement and allowing in rare cases a negative balance.